Add python postgres auth annotations
