+++ backend/python/app/middlewares/auth.py
@@ -2,7 +2,7 @@ from flask import current_app, jsonify, request
 from functools import wraps
 
 from ..services.implementations.auth_service import AuthService
-from ..services.implementations.user_service_mg import UserService
+from ..services.implementations.user_service import UserService
 
 user_service = UserService(current_app.logger)
 auth_service = AuthService(current_app.logger, user_service)
